refactor: Move deleted message handling into its own file
